fix(popup): correct CSS display issue breaking JS search filtering & finalize attachments cleanup
