+++ ui/popup/src/features/attachments/state.js
@@ -1,0 +1,12 @@
+export const attachmentsState = {
+    ticketFiles: [],
+    oppFiles: []
+};
+
+export function clearTicketFiles() {
+    attachmentsState.ticketFiles = [];
+}
+
+export function clearOppFiles() {
+    attachmentsState.oppFiles = [];
+}
